+++ week04/src/apis/auth.ts
@@ -10,21 +10,30 @@ import { axiosInstance } from "./axios";
 export const postSignup = async (
   body: ReqeustSignupDto
 ): Promise<ResponseSignupDto> => {
-  const { data } = await axiosInstance.post<ResponseSignupDto>(
-    `v1/auth/signup`,
-    body
-  );
-
-  return data;
+  try {
+    const { data } = await axiosInstance.post<ResponseSignupDto>(
+      `v1/auth/signup`,
+      body
+    );
+    return data;
+  } catch (e) {
+    console.error(e);
+    throw e;
+  }
 };
 
 // POST v1.auth/signin
 export const postLogin = async (
   body: RequestLoginDto
 ): Promise<ResponseLoginDto> => {
-  const { data } = await axiosInstance.post<ResponseLoginDto>(
-    `v1/auth/signin`,
-    body
-  );
-  return data;
+  try {
+    const { data } = await axiosInstance.post<ResponseLoginDto>(
+      `v1/auth/signin`,
+      body
+    );
+    return data;
+  } catch (e) {
+    console.error(e);
+    throw e;
+  }
 };
